streamline bench trace plumbing
