+++ lib/bench/kv_router/common/progress.rs
@@ -1,0 +1,22 @@
+// SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use indicatif::{ProgressBar, ProgressStyle};
+
+/// Create a styled progress bar, optionally with a known total length.
+pub fn make_progress_bar(total: Option<u64>) -> ProgressBar {
+    let progress = match total {
+        Some(total) => ProgressBar::new(total),
+        None => ProgressBar::no_length(),
+    };
+
+    progress.set_style(
+        ProgressStyle::with_template(
+            "[{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos}/{len} ({eta}) {msg}",
+        )
+        .unwrap()
+        .progress_chars("#>-"),
+    );
+
+    progress
+}
